Much commenting and code cleanup
